Refactored the whole code and added some lacking features

+++ Common/CstringBinarySerializer.h
@@ -9,6 +9,6 @@ public:
 
 	CstringBinarySerializer& operator=(const CstringBinarySerializer&) = delete;
 
-	static void SerializeToOpenedFile(const char* sourceCstring, std::ofstream& file, int size = 0);
+	static void SerializeToOpenedFile(const char* sourceCstring, std::ofstream& file, size_t size = 0ull);
 	static void DeserializeFromOpenedFile(char*& destinationCstringPointer, std::ifstream& file);
 };
